UN-3302 Reorganize service-side code to all be under service
