+++ neuro_san/service/generic/async_agent_service.py
@@ -12,28 +12,27 @@
 
 from typing import Any
 from typing import Dict
-from typing import Iterator
+from typing import Generator
 
-import copy
 import json
 import uuid
 
 from leaf_common.asyncio.asyncio_executor import AsyncioExecutor
 
 from leaf_server_common.server.atomic_counter import AtomicCounter
-from leaf_server_common.server.request_logger import RequestLogger
 
-from neuro_san.internals.graph.registry.agent_network import AgentNetwork
 from neuro_san.internals.interfaces.agent_network_provider import AgentNetworkProvider
+from neuro_san.internals.graph.registry.agent_network import AgentNetwork
 from neuro_san.internals.interfaces.context_type_toolbox_factory import ContextTypeToolboxFactory
 from neuro_san.internals.interfaces.context_type_llm_factory import ContextTypeLlmFactory
 from neuro_san.internals.run_context.factory.master_toolbox_factory import MasterToolboxFactory
 from neuro_san.internals.run_context.factory.master_llm_factory import MasterLlmFactory
-from neuro_san.service.agent_server_logging import AgentServerLogging
-from neuro_san.session.direct_agent_session import DirectAgentSession
+from neuro_san.service.generic.chat_message_converter import ChatMessageConverter
+from neuro_san.service.logging.agent_server_logging import AgentServerLogging
+from neuro_san.service.logging.event_loop_logger import EventLoopLogger
+from neuro_san.session.async_direct_agent_session import AsyncDirectAgentSession
 from neuro_san.session.external_agent_session_factory import ExternalAgentSessionFactory
 from neuro_san.session.session_invocation_context import SessionInvocationContext
-from neuro_san.service.chat_message_converter import ChatMessageConverter
 
 # A list of methods to not log requests for
 # Some of these can be way too chatty
@@ -42,26 +41,22 @@ DO_NOT_LOG_REQUESTS = [
 
 
 # pylint: disable=too-many-instance-attributes
-class AgentService:
+class AsyncAgentService:
     """
-    A base implementation of the Neuro-San Agent Service,
+    A base implementation of the Neuro-San Async Agent Service,
     independent of target transport protocol.
     """
 
     # pylint: disable=too-many-arguments,too-many-positional-arguments
     def __init__(self,
-                 request_logger: RequestLogger,
+                 request_logger: EventLoopLogger,
                  security_cfg: Dict[str, Any],
                  agent_name: str,
                  agent_network_provider: AgentNetworkProvider,
                  server_logging: AgentServerLogging):
         """
-        Set the gRPC interface up for health checking so that the service
-        will be opened to callers when the mesh sees it operational, if this
-        is not done the mesh will treat the service instance as non functional
-
-        :param request_logger: The instance of the RequestLogger that helps
-                    keep track of stats
+        :param request_logger: The instance of the EventLoopLogger that helps
+                        log information from running event loop
         :param security_cfg: A dictionary of parameters used to
                         secure the TLS and the authentication of the gRPC
                         connection.  Supplying this implies use of a secure
@@ -69,7 +64,7 @@ class AgentService:
         :param agent_name: The agent name for the service
         :param agent_network_provider: The AgentNetworkProvider to use for the session.
         :param server_logging: An AgentServerLogging instance initialized so that
-                        spawned asyncrhonous threads can also properly initialize
+                        spawned asynchronous threads can also properly initialize
                         their logging.
         """
         self.request_logger = request_logger
@@ -84,7 +79,7 @@ class AgentService:
         config: Dict[str, Any] = agent_network.get_config()
         self.llm_factory: ContextTypeLlmFactory = MasterLlmFactory.create_llm_factory(config)
         self.toolbox_factory: ContextTypeToolboxFactory = MasterToolboxFactory.create_toolbox_factory(config)
-        # Load once
+        # Load once.
         self.llm_factory.load()
         self.toolbox_factory.load()
 
@@ -94,9 +89,8 @@ class AgentService:
         """
         return self.request_counter.get_count()
 
-    def function(self, request_dict: Dict[str, Any],
-                 request_metadata: Dict[str, Any],
-                 context: Any) \
+    async def function(self, request_dict: Dict[str, Any],
+                       request_metadata: Dict[str, Any]) \
             -> Dict[str, Any]:
         """
         Allows a client to get the outward-facing function for the agent
@@ -104,41 +98,42 @@ class AgentService:
 
         :param request_dict: a FunctionRequest dictionary
         :param request_metadata: request metadata
-        :param context: a service request context object
         :return: a FunctionResponse dictionary
         """
         self.request_counter.increment()
-        request_log = None
+        do_log: bool = "Function" not in DO_NOT_LOG_REQUESTS
         log_marker: str = "function request"
-        service_logging_dict: Dict[str, str] = {
+        metadata: Dict[str, str] = {
             "request_id": f"server-{uuid.uuid4()}"
         }
-        if "Function" not in DO_NOT_LOG_REQUESTS:
-            request_log = self.request_logger.start_request(f"{self.agent_name}.Function",
-                                                            log_marker, context,
-                                                            service_logging_dict)
-
-        # Get the metadata to forward on to another service
-        metadata: Dict[str, str] = copy.copy(service_logging_dict)
         metadata.update(request_metadata)
+        if do_log:
+            self.request_logger.info(
+                metadata,
+                "Received a %s request for %s",
+                f"{self.agent_name}.Function", log_marker)
 
         # Delegate to Direct*Session
         agent_network: AgentNetwork = self.agent_network_provider.get_agent_network()
-        session = DirectAgentSession(agent_network=agent_network,
-                                     invocation_context=None,
-                                     metadata=metadata,
-                                     security_cfg=self.security_cfg)
-        response_dict = session.function(request_dict)
+        session: AsyncDirectAgentSession =\
+            AsyncDirectAgentSession(
+                agent_network=agent_network,
+                invocation_context=None,
+                metadata=metadata,
+                security_cfg=self.security_cfg)
+        response_dict = await session.function(request_dict)
 
-        if request_log is not None:
-            self.request_logger.finish_request(f"{self.agent_name}.Function", log_marker, request_log)
+        if do_log:
+            self.request_logger.info(
+                metadata,
+                "Done with %s request for %s",
+                f"{self.agent_name}.Function", log_marker)
 
         self.request_counter.decrement()
         return response_dict
 
-    def connectivity(self, request_dict: Dict[str, Any],
-                     request_metadata: Dict[str, Any],
-                     context: Any) \
+    async def connectivity(self, request_dict: Dict[str, Any],
+                           request_metadata: Dict[str, Any]) \
             -> Dict[str, Any]:
         """
         Allows a client to get connectivity information for the agent
@@ -146,69 +141,68 @@ class AgentService:
 
         :param request_dict: a ChatRequest dictionary
         :param request_metadata: request metadata
-        :param context: a service request context object
         :return: a ConnectivityResponse dictionary
         """
         self.request_counter.increment()
-        request_log = None
+        do_log: bool = "Connectivity" not in DO_NOT_LOG_REQUESTS
         log_marker: str = "connectivity request"
-        service_logging_dict: Dict[str, str] = {
+        metadata: Dict[str, str] = {
             "request_id": f"server-{uuid.uuid4()}"
         }
-        if "Connectivity" not in DO_NOT_LOG_REQUESTS:
-            request_log = self.request_logger.start_request(f"{self.agent_name}.Connectivity",
-                                                            log_marker, context,
-                                                            service_logging_dict)
-
-        # Get the metadata to forward on to another service
-        metadata: Dict[str, str] = copy.copy(service_logging_dict)
         metadata.update(request_metadata)
+
+        if do_log:
+            self.request_logger.info(
+                metadata,
+                "Received a %s request for %s",
+                f"{self.agent_name}.Connectivity", log_marker)
 
         # Delegate to Direct*Session
         agent_network: AgentNetwork = self.agent_network_provider.get_agent_network()
-        session = DirectAgentSession(agent_network=agent_network,
-                                     invocation_context=None,
-                                     metadata=metadata,
-                                     security_cfg=self.security_cfg)
-        response_dict = session.connectivity(request_dict)
+        session: AsyncDirectAgentSession =\
+            AsyncDirectAgentSession(
+                agent_network=agent_network,
+                invocation_context=None,
+                metadata=metadata,
+                security_cfg=self.security_cfg)
+        response_dict = await session.connectivity(request_dict)
 
-        if request_log is not None:
-            self.request_logger.finish_request(f"{self.agent_name}.Connectivity", log_marker, request_log)
+        if do_log:
+            self.request_logger.info(
+                metadata,
+                "Done with %s request for %s",
+                f"{self.agent_name}.Connectivity", log_marker)
 
         self.request_counter.decrement()
         return response_dict
 
     # pylint: disable=too-many-locals
-    def streaming_chat(self, request_dict: Dict[str, Any],
-                       request_metadata: Dict[str, Any],
-                       context: Any) \
-            -> Iterator[Dict[str, Any]]:
+    async def streaming_chat(self, request_dict: Dict[str, Any],
+                             request_metadata: Dict[str, Any]) \
+            -> Generator[Dict[str, Any], None, None]:
         """
         Initiates or continues the agent chat with the session_id
         context in the request.
 
         :param request_dict: a ChatRequest dictionary
         :param request_metadata: request metadata
-        :param context: a service request context object
         :return: an iterator for (eventually) returned responses dictionaries
         """
         self.request_counter.increment()
-        request_log = None
         user_text: str = request_dict.get("user_message", {}).get("text", "")
+        do_log: bool = "StreamingChat" not in DO_NOT_LOG_REQUESTS
+
         log_marker = f"'{user_text}'"
-        service_logging_dict: Dict[str, str] = {
+        metadata: Dict[str, str] = {
             "request_id": f"server-{uuid.uuid4()}"
         }
-        if "StreamingChat" not in DO_NOT_LOG_REQUESTS:
-            request_log = self.request_logger.start_request(f"{self.agent_name}.StreamingChat",
-                                                            log_marker, context,
-                                                            service_logging_dict)
-
-        # Get the metadata to forward on to another service
-        metadata: Dict[str, str] = copy.copy(service_logging_dict)
         metadata.update(request_metadata)
-        if metadata.get("request_id") is None:
-            metadata["request_id"] = service_logging_dict.get("request_id")
+
+        if do_log:
+            self.request_logger.info(
+                metadata,
+                "Received a %s request for %s",
+                f"{self.agent_name}.StreamingChat", log_marker)
 
         # Prepare
         factory = ExternalAgentSessionFactory(use_direct=False)
@@ -222,19 +216,21 @@ class AgentService:
 
         # Delegate to Direct*Session
         agent_network: AgentNetwork = self.agent_network_provider.get_agent_network()
-        session = DirectAgentSession(agent_network=agent_network,
-                                     invocation_context=invocation_context,
-                                     metadata=metadata,
-                                     security_cfg=self.security_cfg)
-        # Get our args in order to pass to grpc-free session level
-        response_dict_iterator: Iterator[Dict[str, Any]] = session.streaming_chat(request_dict)
+        session: AsyncDirectAgentSession =\
+            AsyncDirectAgentSession(
+                agent_network=agent_network,
+                invocation_context=invocation_context,
+                metadata=metadata,
+                security_cfg=self.security_cfg)
+        # Get our args in order to pass to transport-agnostic session level
+        response_dict_generator: Generator[Dict[str, Any], None, None] = session.streaming_chat(request_dict)
 
         # See if we want to put the request dict in the response
         chat_filter_dict: Dict[str, Any] = {}
         chat_filter_dict = request_dict.get("chat_filter", chat_filter_dict)
         chat_filter_type: str = chat_filter_dict.get("chat_filter_type", "MINIMAL")
 
-        for response_dict in response_dict_iterator:
+        async for response_dict in response_dict_generator:
             # Prepare chat message for output:
             response_dict = ChatMessageConverter().to_dict(response_dict)
             # Do not return the request when the filter is MINIMAL
@@ -247,11 +243,14 @@ class AgentService:
 
         # Iterator has finally signaled that there are no more responses to be had.
         # Log that we are done.
-        if request_log is not None:
+        if do_log:
             reporting: str = None
             if request_reporting is not None:
                 reporting = json.dumps(request_reporting, indent=4, sort_keys=True)
-            request_log.metrics("Request reporting: %s", reporting)
-            self.request_logger.finish_request(f"{self.agent_name}.StreamingChat", log_marker, request_log)
+            self.request_logger.info(metadata, "Request reporting: %s", reporting)
+            self.request_logger.info(
+                metadata,
+                "Done with %s request for %s",
+                f"{self.agent_name}.StreamingChat", log_marker)
 
         self.request_counter.decrement()
