+++ neuro_san/service/http/handlers/function_handler.py
@@ -14,8 +14,8 @@ See class comment for details
 """
 from typing import Any, Dict
 
-from neuro_san.http_sidecar.handlers.base_request_handler import BaseRequestHandler
-from neuro_san.service.async_agent_service import AsyncAgentService
+from neuro_san.service.generic.async_agent_service import AsyncAgentService
+from neuro_san.service.http.handlers.base_request_handler import BaseRequestHandler
 
 
 class FunctionHandler(BaseRequestHandler):
