+++ neuro_san/service/http/handlers/health_check_handler.py
@@ -22,7 +22,7 @@ from importlib.metadata import version as library_version
 from importlib.metadata import PackageNotFoundError
 
 from tornado.web import RequestHandler
-from neuro_san.http_sidecar.logging.http_logger import HttpLogger
+from neuro_san.service.http.logging.http_logger import HttpLogger
 
 
 class HealthCheckHandler(RequestHandler):
