+++ neuro_san/service/http/handlers/openapi_publish_handler.py
@@ -12,35 +12,34 @@
 """
 See class comment for details
 """
-from typing import Any, Dict
+from typing import Any
+from typing import Dict
 
-from neuro_san.http_sidecar.handlers.base_request_handler import BaseRequestHandler
-from neuro_san.interfaces.concierge_session import ConciergeSession
-from neuro_san.session.direct_concierge_session import DirectConciergeSession
+import json
+
+from neuro_san.service.http.handlers.base_request_handler import BaseRequestHandler
 
 
-class ConciergeHandler(BaseRequestHandler):
+class OpenApiPublishHandler(BaseRequestHandler):
     """
-    Handler class for neuro-san "concierge" API call.
+    Handler class for neuro-san OpenAPI service spec publishing"concierge" API call.
     """
 
     def get(self):
         """
-        Implementation of GET request handler for "concierge" API call.
+        Implementation of GET request handler
+        for "publish my OpenAPI specification document" call.
         """
         metadata: Dict[str, Any] = self.get_metadata()
-        self.application.start_client_request(metadata, "/api/v1/list")
+        self.application.start_client_request(metadata, "/api/v1/docs")
         try:
-            data: Dict[str, Any] = {}
-            session: ConciergeSession = DirectConciergeSession(metadata=metadata)
-            result_dict: Dict[str, Any] = session.list(data)
-
-            # Return gRPC response to the HTTP client
+            with open(self.openapi_service_spec_path, "r", encoding='utf-8') as f_out:
+                result_dict: Dict[str, Any] = json.load(f_out)
+            # Return json data to the HTTP client
             self.set_header("Content-Type", "application/json")
             self.write(result_dict)
-
         except Exception as exc:  # pylint: disable=broad-exception-caught
             self.process_exception(exc)
         finally:
             self.do_finish()
-            self.application.finish_client_request(metadata, "/api/v1/list")
+            self.application.finish_client_request(metadata, "/api/v1/docs")
