+++ neuro_san/service/http/handlers/streaming_chat_handler.py
@@ -12,11 +12,14 @@
 """
 See class comment for details
 """
-from typing import Any, Dict, Generator
+from typing import Any
+from typing import Dict
+from typing import Generator
+
 import json
 
-from neuro_san.http_sidecar.handlers.base_request_handler import BaseRequestHandler
-from neuro_san.service.async_agent_service import AsyncAgentService
+from neuro_san.service.generic.async_agent_service import AsyncAgentService
+from neuro_san.service.http.handlers.base_request_handler import BaseRequestHandler
 
 
 class StreamingChatHandler(BaseRequestHandler):
