+++ neuro_san/service/http/logging/http_logger.py
@@ -19,8 +19,8 @@ from typing import Any, Dict, Sequence
 
 from leaf_server_common.logging.logging_setup import setup_logging
 
-from neuro_san.http_sidecar.logging.event_loop_logger import EventLoopLogger
-from neuro_san.http_sidecar.logging.log_context_filter import LogContextFilter
+from neuro_san.service.http.logging.log_context_filter import LogContextFilter
+from neuro_san.service.logging.event_loop_logger import EventLoopLogger
 
 
 class HttpLogger(EventLoopLogger):
